move workersController to root directory
